Update testutils and test cases.
1. change min to lowest.
2. remove quant_params cause it only change shape of tensor.

ONE-DCO-1.0-Signed-off-by: KiDeuk Bang <rrstrous@nate.com>

+++ compiler/luci-interpreter/src/loader/KernelBuilder.cpp
@@ -466,7 +466,7 @@ std::unique_ptr<Kernel> KernelBuilder::visit(const luci::CircleSqueeze *node)
   SqueezeParams params{};
   assert(node->squeeze_dims().size() <= 4);
   params.squeeze_dims_count = node->squeeze_dims().size();
-  for (size_t i = 0; i < params.squeeze_dims_count; i++)
+  for (int8_t i = 0; i < params.squeeze_dims_count; i++)
   {
     params.squeeze_dims[i] = node->squeeze_dims().at(i);
   }
